requirements.txt
django-pint